requirements.txt
Django==2.0.4
django-braces==1.12.0
django-filter==1.1.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
pytz==2018.4
